docker/applications/requirements.txt
aioredis
aiomysql
pystrix
asyncio
